Support namespaces in the Connect injector
This adds namespace config options for registering injected
services into a single namespace as well as mirroring k8s
namespaces in Consul with an optional prefix.

It adds functionality to check for Consul namespace existence
and create new namespaces. If mirroring namespaces with ACLs
enabled, binding rules are also checked and created if necessary.

Service and proxy registration as well as service-defaults
have been updated to be namespace aware.

Open issues:
- Meta on namespaces is not yet supported in OSS api
- `consul config write` does not take a `-ns` flag yet
- `consul logout` does not take a `-ns` flag yet
- `service-defaults` may or may not need a namespace field
- Consul api import will need to be updated to a released version